refactor: delete old test file and split into four separate files for modularity
